AIML Projects/requirements.txt
Pillow
google-generativeai
matplotlib
mediapipe
numpy
opencv-python
pandas
psutil
python-dotenv
pyttsx3
pywhatkit
reportlab
requests
scikit-learn
seaborn
SpeechRecognition
streamlit
textblob
fastapi
uvicorn
